Rename dotnet_sdk_command
